+++ addon/plugin.program.parsec-launcher/lib/addonlang.py
@@ -1,0 +1,38 @@
+
+
+import xbmcaddon
+
+addon = xbmcaddon.Addon(id='plugin.program.parsec-launcher')
+
+# Language strings
+LANG_TITLE_CONNECT_PARSEC=addon.getLocalizedString(30020).encode('utf-8')
+LANG_TITLE_MANAGE_COMPUTERS=addon.getLocalizedString(30021).encode('utf-8')
+LANG_TITLE_MANAGE=addon.getLocalizedString(30022).encode('utf-8')
+LANG_TITLE_MANAGE_COMPUTER_SWITCH_ON=addon.getLocalizedString(30023).encode('utf-8')
+LANG_TITLE_MANAGE_COMPUTER_SWITCH_OFF=addon.getLocalizedString(30024).encode('utf-8')
+LANG_TITLE_MANAGE_COMPUTER_IS_PENDING=addon.getLocalizedString(30025).encode('utf-8')
+LANG_PARSEC=addon.getLocalizedString(30026).encode('utf-8')
+LANG_TITLE_STATUS=addon.getLocalizedString(30027).encode('utf-8')
+LANG_TITLE_CREATED=addon.getLocalizedString(30028).encode('utf-8')
+LANG_TITLE_LAST_UPDATED=addon.getLocalizedString(30029).encode('utf-8')
+LANG_TITLE_PROVIDER=addon.getLocalizedString(30030).encode('utf-8')
+LANG_TITLE_MACHINE_TYPE=addon.getLocalizedString(30031).encode('utf-8')
+LANG_TITLE_REGION=addon.getLocalizedString(30032).encode('utf-8')
+LANG_TITLE_CREDIT=addon.getLocalizedString(30033).encode('utf-8')
+LANG_TITLE_PLAYTIME=addon.getLocalizedString(30034).encode('utf-8')
+LANG_TITLE_HOURS = addon.getLocalizedString(30035).encode('utf-8')
+LANG_TITLE_DOLLAR = addon.getLocalizedString(30036).encode('utf-8')
+LANG_TITLE_COMPUTER_INFO = addon.getLocalizedString(30037).encode('utf-8')
+LANG_TITLE_USER_INFO = addon.getLocalizedString(30038).encode('utf-8')
+LANG_TITLE_USERNAME = addon.getLocalizedString(30039).encode('utf-8')
+LANG_MESSAGE_NO_CREDENTIALS=addon.getLocalizedString(30040).encode('utf-8')
+LANG_QUESTION_TO_SETTINGS=addon.getLocalizedString(30041).encode('utf-8')
+LANG_MESSAGE_WRONG_CREDENTIALS=addon.getLocalizedString(30042).encode('utf-8')
+LANG_MESSAGE_COMPUTER_OFF=addon.getLocalizedString(30043).encode('utf-8')
+LANG_QUESTION_QUESTION_SWITCH_ON=addon.getLocalizedString(30044).encode('utf-8')
+LANG_MESSAGE_COMPUTER_ON=addon.getLocalizedString(30045).encode('utf-8')
+LANG_MESSAGE_BACKGROUND_TASK_BREAK=addon.getLocalizedString(30046).encode('utf-8')
+
+LANG_TITLE_MANAGE_COMPUTER_MESSAGE_PENDING=addon.getLocalizedString(30200).encode('utf-8')
+LANG_TITLE_MANAGE_COMPUTER_MESSAGE_ON=addon.getLocalizedString(30201).encode('utf-8')
+LANG_TITLE_MANAGE_COMPUTER_MESSAGE_OFF=addon.getLocalizedString(30202).encode('utf-8')
